+++ src/onegov.agency/onegov/agency/forms/__init__.py
@@ -1,9 +1,11 @@
 from onegov.agency.forms.agency import ExtendedAgencyForm
 from onegov.agency.forms.membership import MembershipForm
+from onegov.agency.forms.mutation import MutationForm
 
 
 __all__ = (
     'ExtendedAgencyForm',
     'ExtendedPersonForm',
     'MembershipForm',
+    'MutationForm',
 )